モジュール化とウミガメのスープ機能の実装
