+++ .gitignore
@@ -1,3 +1,3 @@
-venv
+*venv
 .env
 __pycache__
