+++ .vscode/settings.json
@@ -1,3 +1,12 @@
 {
-  "files.autoSave": "afterDelay"
+  "files.autoSave": "afterDelay",
+  "python.testing.unittestArgs": [
+    "-v",
+    "-s",
+    "./venv",
+    "-p",
+    "*_test.py"
+  ],
+  "python.testing.pytestEnabled": false,
+  "python.testing.unittestEnabled": true
 }
