+++ pyproject.toml
@@ -3,8 +3,7 @@ name = "discordbot"
 version = "0.1.0"
 description = ""
 authors = ["yootako <yootako03@gamil.com>"]
-# readme = "README.md" 
-package-mode = false
+# readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.12"
@@ -12,6 +11,7 @@ requests = "^2.32.3"
 yt-dlp = "^2024.11.4"
 python-dotenv = "^1.0.1"
 discord-py = {extras = ["voice"], version = "^2.4.0"}
+google-genai = "^1.14.0"
 
 
 [build-system]
