+++ src/__init__.py
@@ -1,0 +1,1 @@
+# umigame モジュールの初期化ファイル
