+++ src/config.py
@@ -10,3 +10,4 @@ import logging
 logger = logging.getLogger(__name__)
 
 DISCORD_TOKEN = os.getenv('DISCORD_TOKEN')
+GEMINI_API_KEY = os.getenv('Gemini_API_KEY')
