+++ src/umigame/README.md
@@ -1,0 +1,21 @@
+# うみがめのスープ
+Discordのコマンドを受け取ってチャットに自動応答する
+GeminiAPIを使って自動応答する
+
+1. お題の生成
+    - うみがめのスープの問題を生成する
+    - 例: 「ある男がレストランでうみがめのスープを注文した。食べていると、突然、男は泣き出した。なぜ？」
+    - 生成したお題は変数に格納する
+
+2. 質問の受け答え
+    - 質問を受け取る
+    - 質問に対して「はい」「いいえ」「わからない」で答える
+    - 正解した場合は「正解！」と返す
+    - 例: 「その男はうみがめのスープを食べたことがある？」→「はい」
+
+3. 正解した場合
+    - 正解した場合は「正解！」と返す
+    - 例: 「その男はうみがめのスープを食べたことがある？」→「はい」
+    - 正解の理由を説明する
+    - 例: 「その男はうみがめのスープを食べたことがあるから、泣いた」
+
