+++ src/yomiage/__init__.py
@@ -1,0 +1,1 @@
+# yomiage モジュールの初期化ファイル
